+++ src/main/java/com/example/review/controller/HospitalController.java
@@ -41,7 +41,7 @@ public class HospitalController {
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable Integer id, Model model, Pageable pageable) {
+    public String show(@PathVariable Long id, Model model, Pageable pageable) {
         Optional<Hospital> hospital = hospitalRepository.findById(id);
         Page<Review> reviews = reviewRepository.findByHospitalId(id, pageable);
         log.info("review cnt:{} {}", reviews.getSize(), reviews);
